Move analyzer to package
