+++ Tapanga.CommandLine.Analyzers/Tapanga.CommandLine.Analyzers.csproj
@@ -1,0 +1,21 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <ImplicitUsings>disable</ImplicitUsings>
+    <Nullable>disable</Nullable>
+    <!-- Do not include the generator as a lib dependency -->
+    <IncludeBuildOutput>false</IncludeBuildOutput>
+    <Configurations>Debug;Release;ReleaseLocal</Configurations>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="SourceGeneratorsKit" Version="0.0.1" PrivateAssets="all" GeneratePathProperty="true" />
+  </ItemGroup>
+  
+  <ItemGroup>
+    <None Include="$(OutputPath)\$(AssemblyName).dll" Pack="true" PackagePath="analyzers/dotnet/cs" Visible="false" />
+    <None Include="$(PkgSourceGeneratorsKit)\lib\netstandard2.0\*.dll" Pack="true" PackagePath="analyzers/dotnet/cs" Visible="false" />
+  </ItemGroup>
+
+</Project>
