--- Tapanga.CommandLine/SourceGenerators/EntryPointGenerator.cs
@@ -1,51 +1,0 @@
-﻿using Microsoft.CodeAnalysis;
-using SourceGeneratorsKit;
-using Tapanga.Plugin;
-
-namespace Tapanga.CommandLine.SourceGenerators;
-
-/// <summary>
-/// Generates the entry point for a tapanga plugin tool.
-/// </summary>
-[Generator]
-public class EntryPointGenerator : ISourceGenerator
-{
-    ClassesWithInterfacesReceiver interfacesReceiver = new ClassesWithInterfacesReceiver(nameof(IGeneratorProvider));
-
-    /// <inheritdoc />
-    public void Initialize(GeneratorInitializationContext context)
-    {
-        context.RegisterForSyntaxNotifications(() => interfacesReceiver);
-    }
-
-    /// <inheritdoc />
-    public void Execute(GeneratorExecutionContext context)
-    {
-        if (this.interfacesReceiver.Classes.SingleOrDefault() is INamedTypeSymbol providerSymbol)
-        {
-            string? description = providerSymbol.ToDisplayString();
-
-            if (providerSymbol.FindAttribute("Description") is AttributeData attribute)
-            {
-                description = (string)attribute.ConstructorArguments[0].Value!;
-            }
-
-            string source = $@"
-
-namespace {providerSymbol.ToDisplayString()};
-
-public class Program
-{{
-    public static async Task<int> Main(string[] args)
-    {{
-        var provider = new {providerSymbol.ContainingNamespace.ToDisplayString()}.{providerSymbol.Name}();
-        var factories = provider.{nameof(IGeneratorProvider.GetGeneratorFactories)}();
-        await new Tapanga.CommandLine.Runner(""{description}"", factories).InvokeAsync(args);
-    }}
-}}
-";
-
-            context.AddSource("Program", source);
-        }
-    }
-}
