+++ Tapanga.sln
@@ -9,6 +9,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tapanga.Plugin", "Tapanga.P
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tapanga.CommandLine", "Tapanga.CommandLine\Tapanga.CommandLine.csproj", "{21EE6185-E535-476E-B537-DBB444002C36}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tapanga.CommandLine.Analyzers", "Tapanga.CommandLine.Analyzers\Tapanga.CommandLine.Analyzers.csproj", "{0EAB5843-8CB0-40CD-9E6C-02E92098A43D}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -34,6 +36,12 @@ Global
 		{21EE6185-E535-476E-B537-DBB444002C36}.Release|Any CPU.Build.0 = Release|Any CPU
 		{21EE6185-E535-476E-B537-DBB444002C36}.ReleaseLocal|Any CPU.ActiveCfg = ReleaseLocal|Any CPU
 		{21EE6185-E535-476E-B537-DBB444002C36}.ReleaseLocal|Any CPU.Build.0 = ReleaseLocal|Any CPU
+		{0EAB5843-8CB0-40CD-9E6C-02E92098A43D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0EAB5843-8CB0-40CD-9E6C-02E92098A43D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0EAB5843-8CB0-40CD-9E6C-02E92098A43D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0EAB5843-8CB0-40CD-9E6C-02E92098A43D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0EAB5843-8CB0-40CD-9E6C-02E92098A43D}.ReleaseLocal|Any CPU.ActiveCfg = ReleaseLocal|Any CPU
+		{0EAB5843-8CB0-40CD-9E6C-02E92098A43D}.ReleaseLocal|Any CPU.Build.0 = ReleaseLocal|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
